I went ahead and added simple support for time windowed aggregations. Metrics are read, unmarshaled, filtered, group by metric name (no dimensions yet), simple filter metric name to aggregated metric name.
The timestamp of a metric is used to evaluated the event timed window that the metric belongs in. metric.timestamp / windowSize.

There is a map of maps. The first level of the map, the key is an int which represents the event time window that the metrics should be placed in.

the second level of the map is the windowed aggregations for a specific event time window. This is terminology adopted from Kafka Streams. The aggregations are just sums of the values based on the aggregated/target metric name. Nothing fancy right now.

On windowSize time periods the metrics are output to a log for the last window.

I was able to measure 50K metrics per second injest rate, with three aggregation specifications, and three separate metrics being aggregated on my MacBook with around 50-70% cpu utilization. There were 100 different metrics in total, with three of them being aggregated on, out of the 100.

Also added a test publisher.

windowsSize is adjustable to any value.