Added support for abitrary windowLag, not just one back, and delete just published timeWindow from the map
WIndow lag can now be greater than window size allowing for publishing aggregations for several time windows in the past. For example, it might be desirable to process aggregations several time windows back when the window size is very small, such as a minute, to allow enough time for delays in the system due to the Agent and other areas.